+++ src/main/java/org/nrg/containers/model/server/docker/DockerServerBase.java
@@ -19,11 +19,12 @@ public abstract class DockerServerBase {
     @JsonIgnore public abstract Date lastEventCheckTime();
     @Nullable @JsonProperty("path-translation-xnat-prefix") public abstract String pathTranslationXnatPrefix();
     @Nullable @JsonProperty("path-translation-docker-prefix") public abstract String pathTranslationDockerPrefix();
+    @JsonProperty("pull-images-on-xnat-init") public abstract Boolean pullImagesOnXnatInit();
 
     @AutoValue
     @JsonInclude(JsonInclude.Include.ALWAYS)
     public abstract static class DockerServer extends DockerServerBase {
-        public static final DockerServer DEFAULT_SOCKET = DockerServer.create(0L, "Local socket", "unix:///var/run/docker.sock", null, false, null, null);
+        public static final DockerServer DEFAULT_SOCKET = DockerServer.create("Local socket", "unix:///var/run/docker.sock");
 
         @JsonCreator
         public static DockerServer create(@JsonProperty("id") final Long id,
@@ -32,8 +33,14 @@ public abstract class DockerServerBase {
                                           @JsonProperty("cert-path") final String certPath,
                                           @JsonProperty("swarm-mode") final Boolean swarmMode,
                                           @JsonProperty("path-translation-xnat-prefix") final String pathTranslationXnatPrefix,
-                                          @JsonProperty("path-translation-docker-prefix") final String pathTranslationDockerPrefix) {
-            return create(id, name, host, certPath, swarmMode, null, pathTranslationXnatPrefix, pathTranslationDockerPrefix);
+                                          @JsonProperty("path-translation-docker-prefix") final String pathTranslationDockerPrefix,
+                                          @JsonProperty("pull-images-on-xnat-init") final Boolean pullImagesOnXnatInit) {
+            return create(id, name, host, certPath, swarmMode, null, pathTranslationXnatPrefix, pathTranslationDockerPrefix, pullImagesOnXnatInit);
+        }
+
+        public static DockerServer create(final String name,
+                                          final String host) {
+            return create(0L, name, host, null, false, null, null, null, null);
         }
 
         public static DockerServer create(final Long id,
@@ -43,7 +50,8 @@ public abstract class DockerServerBase {
                                           final Boolean swarmMode,
                                           final Date lastEventCheckTime,
                                           final String pathTranslationXnatPrefix,
-                                          final String pathTranslationDockerPrefix) {
+                                          final String pathTranslationDockerPrefix,
+                                          final Boolean pullImagesOnXnatInit) {
             return new AutoValue_DockerServerBase_DockerServer(
                     id == null ? 0L : id,
                     StringUtils.isBlank(name) ? host : name,
@@ -52,11 +60,13 @@ public abstract class DockerServerBase {
                     swarmMode != null && swarmMode,
                     lastEventCheckTime != null ? lastEventCheckTime : new Date(),
                     pathTranslationXnatPrefix,
-                    pathTranslationDockerPrefix
+                    pathTranslationDockerPrefix,
+                    pullImagesOnXnatInit != null && pullImagesOnXnatInit
             );
         }
 
         public static DockerServer create(final DockerServerEntity dockerServerEntity) {
+            final Boolean pullImagesOnXnatInit = dockerServerEntity.getPullImagesOnXnatInit();
             return create(
                     dockerServerEntity.getId(),
                     dockerServerEntity.getName(),
@@ -65,7 +75,8 @@ public abstract class DockerServerBase {
                     dockerServerEntity.getSwarmMode(),
                     dockerServerEntity.getLastEventCheckTime(),
                     dockerServerEntity.getPathTranslationXnatPrefix(),
-                    dockerServerEntity.getPathTranslationDockerPrefix()
+                    dockerServerEntity.getPathTranslationDockerPrefix(),
+                    pullImagesOnXnatInit == null ? false : pullImagesOnXnatInit
             );
         }
 
@@ -79,7 +90,8 @@ public abstract class DockerServerBase {
                     false,
                     dockerServerPrefsBean.getLastEventCheckTime(),
                     null,
-                    null);
+                    null,
+                    false);
         }
 
         public DockerServer updateEventCheckTime(final Date newLastEventCheckTime) {
@@ -93,7 +105,8 @@ public abstract class DockerServerBase {
                             this.swarmMode(),
                             newLastEventCheckTime,
                             this.pathTranslationXnatPrefix(),
-                            this.pathTranslationDockerPrefix()
+                            this.pathTranslationDockerPrefix(),
+                            this.pullImagesOnXnatInit()
                     );
         }
     }
@@ -111,8 +124,10 @@ public abstract class DockerServerBase {
                                                   @JsonProperty("swarm-mode") final Boolean swarmMode,
                                                   @JsonProperty("path-translation-xnat-prefix") final String pathTranslationXnatPrefix,
                                                   @JsonProperty("path-translation-docker-prefix") final String pathTranslationDockerPrefix,
+                                                  @JsonProperty("pull-images-on-xnat-init") final Boolean pullImagesOnXnatInit,
                                                   @JsonProperty("ping") final Boolean ping) {
-            return create(id == null ? 0L : id, name, host, certPath, swarmMode, new Date(0), pathTranslationXnatPrefix, pathTranslationDockerPrefix, ping);
+            return create(id == null ? 0L : id, name, host, certPath, swarmMode, new Date(0),
+                    pathTranslationXnatPrefix, pathTranslationDockerPrefix, pullImagesOnXnatInit, ping);
         }
 
         public static DockerServerWithPing create(final Long id,
@@ -123,6 +138,7 @@ public abstract class DockerServerBase {
                                                   final Date lastEventCheckTime,
                                                   final String pathTranslationXnatPrefix,
                                                   final String pathTranslationDockerPrefix,
+                                                  final Boolean pullImagesOnXnatInit,
                                                   final Boolean ping) {
             return new AutoValue_DockerServerBase_DockerServerWithPing(
                     id == null ? 0L : id,
@@ -133,6 +149,7 @@ public abstract class DockerServerBase {
                     lastEventCheckTime != null ? lastEventCheckTime : new Date(0),
                     pathTranslationXnatPrefix,
                     pathTranslationDockerPrefix,
+                    pullImagesOnXnatInit != null && pullImagesOnXnatInit,
                     ping != null && ping);
         }
 
@@ -147,6 +164,7 @@ public abstract class DockerServerBase {
                     dockerServer.lastEventCheckTime(),
                     dockerServer.pathTranslationXnatPrefix(),
                     dockerServer.pathTranslationDockerPrefix(),
+                    dockerServer.pullImagesOnXnatInit(),
                     ping
             );
         }